+++ packages/core/CHANGELOG.md
@@ -3,6 +3,79 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [10.4.0](https://github.com/furystack/furystack/compare/@furystack/core@10.3.8...@furystack/core@10.4.0) (2021-11-30)
+
+
+### 🚀 What's new
+
+* **@furystack/core:** Advanced filters for InMemory store ([bc89ecc](https://github.com/furystack/furystack/commit/bc89ecc49781d9f6ec1b772662e00d3467b15227))
+
+
+
+
+### [10.3.8](https://github.com/furystack/furystack/compare/@furystack/core@10.3.7...@furystack/core@10.3.8) (2021-11-29)
+
+**Note:** Version bump only for package @furystack/core
+
+
+
+
+
+
+### [10.3.7](https://github.com/furystack/furystack/compare/@furystack/core@10.3.6...@furystack/core@10.3.7) (2021-11-19)
+
+**Note:** Version bump only for package @furystack/core
+
+
+
+
+
+
+### [10.3.6](https://github.com/furystack/furystack/compare/@furystack/core@10.3.5...@furystack/core@10.3.6) (2021-11-09)
+
+**Note:** Version bump only for package @furystack/core
+
+
+
+
+
+
+### [10.3.5](https://github.com/furystack/furystack/compare/@furystack/core@10.3.4...@furystack/core@10.3.5) (2021-10-15)
+
+**Note:** Version bump only for package @furystack/core
+
+
+
+
+
+
+### [10.3.4](https://github.com/furystack/furystack/compare/@furystack/core@10.3.3...@furystack/core@10.3.4) (2021-10-05)
+
+**Note:** Version bump only for package @furystack/core
+
+
+
+
+
+
+### [10.3.3](https://github.com/furystack/furystack/compare/@furystack/core@10.3.2...@furystack/core@10.3.3) (2021-09-16)
+
+**Note:** Version bump only for package @furystack/core
+
+
+
+
+
+
+### [10.3.2](https://github.com/furystack/furystack/compare/@furystack/core@10.3.1...@furystack/core@10.3.2) (2021-08-27)
+
+**Note:** Version bump only for package @furystack/core
+
+
+
+
+
+
 ### [10.3.1](https://github.com/furystack/furystack/compare/@furystack/core@10.3.0...@furystack/core@10.3.1) (2021-08-19)
 
 **Note:** Version bump only for package @furystack/core