+++ packages/repository/CHANGELOG.md
@@ -3,6 +3,78 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [5.0.24](https://github.com/furystack/furystack/compare/@furystack/repository@5.0.23...@furystack/repository@5.0.24) (2021-11-30)
+
+**Note:** Version bump only for package @furystack/repository
+
+
+
+
+
+
+### [5.0.23](https://github.com/furystack/furystack/compare/@furystack/repository@5.0.22...@furystack/repository@5.0.23) (2021-11-29)
+
+**Note:** Version bump only for package @furystack/repository
+
+
+
+
+
+
+### [5.0.22](https://github.com/furystack/furystack/compare/@furystack/repository@5.0.21...@furystack/repository@5.0.22) (2021-11-19)
+
+**Note:** Version bump only for package @furystack/repository
+
+
+
+
+
+
+### [5.0.21](https://github.com/furystack/furystack/compare/@furystack/repository@5.0.20...@furystack/repository@5.0.21) (2021-11-09)
+
+**Note:** Version bump only for package @furystack/repository
+
+
+
+
+
+
+### [5.0.20](https://github.com/furystack/furystack/compare/@furystack/repository@5.0.19...@furystack/repository@5.0.20) (2021-10-15)
+
+**Note:** Version bump only for package @furystack/repository
+
+
+
+
+
+
+### [5.0.19](https://github.com/furystack/furystack/compare/@furystack/repository@5.0.18...@furystack/repository@5.0.19) (2021-10-05)
+
+**Note:** Version bump only for package @furystack/repository
+
+
+
+
+
+
+### [5.0.18](https://github.com/furystack/furystack/compare/@furystack/repository@5.0.17...@furystack/repository@5.0.18) (2021-09-16)
+
+**Note:** Version bump only for package @furystack/repository
+
+
+
+
+
+
+### [5.0.17](https://github.com/furystack/furystack/compare/@furystack/repository@5.0.16...@furystack/repository@5.0.17) (2021-08-27)
+
+**Note:** Version bump only for package @furystack/repository
+
+
+
+
+
+
 ### [5.0.16](https://github.com/furystack/furystack/compare/@furystack/repository@5.0.15...@furystack/repository@5.0.16) (2021-08-19)
 
 **Note:** Version bump only for package @furystack/repository