+++ packages/rest-client-got/CHANGELOG.md
@@ -3,6 +3,78 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [1.0.33](https://github.com/furystack/furystack/compare/@furystack/rest-client-got@1.0.32...@furystack/rest-client-got@1.0.33) (2021-11-30)
+
+**Note:** Version bump only for package @furystack/rest-client-got
+
+
+
+
+
+
+### [1.0.32](https://github.com/furystack/furystack/compare/@furystack/rest-client-got@1.0.31...@furystack/rest-client-got@1.0.32) (2021-11-29)
+
+**Note:** Version bump only for package @furystack/rest-client-got
+
+
+
+
+
+
+### [1.0.31](https://github.com/furystack/furystack/compare/@furystack/rest-client-got@1.0.30...@furystack/rest-client-got@1.0.31) (2021-11-19)
+
+**Note:** Version bump only for package @furystack/rest-client-got
+
+
+
+
+
+
+### [1.0.30](https://github.com/furystack/furystack/compare/@furystack/rest-client-got@1.0.29...@furystack/rest-client-got@1.0.30) (2021-11-09)
+
+**Note:** Version bump only for package @furystack/rest-client-got
+
+
+
+
+
+
+### [1.0.29](https://github.com/furystack/furystack/compare/@furystack/rest-client-got@1.0.28...@furystack/rest-client-got@1.0.29) (2021-10-15)
+
+**Note:** Version bump only for package @furystack/rest-client-got
+
+
+
+
+
+
+### [1.0.28](https://github.com/furystack/furystack/compare/@furystack/rest-client-got@1.0.27...@furystack/rest-client-got@1.0.28) (2021-10-05)
+
+**Note:** Version bump only for package @furystack/rest-client-got
+
+
+
+
+
+
+### [1.0.27](https://github.com/furystack/furystack/compare/@furystack/rest-client-got@1.0.26...@furystack/rest-client-got@1.0.27) (2021-09-16)
+
+**Note:** Version bump only for package @furystack/rest-client-got
+
+
+
+
+
+
+### [1.0.26](https://github.com/furystack/furystack/compare/@furystack/rest-client-got@1.0.25...@furystack/rest-client-got@1.0.26) (2021-08-27)
+
+**Note:** Version bump only for package @furystack/rest-client-got
+
+
+
+
+
+
 ### [1.0.25](https://github.com/furystack/furystack/compare/@furystack/rest-client-got@1.0.24...@furystack/rest-client-got@1.0.25) (2021-08-19)
 
 **Note:** Version bump only for package @furystack/rest-client-got