+++ packages/rest-service/CHANGELOG.md
@@ -3,6 +3,97 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [4.1.4](https://github.com/furystack/furystack/compare/@furystack/rest-service@4.1.3...@furystack/rest-service@4.1.4) (2021-11-30)
+
+**Note:** Version bump only for package @furystack/rest-service
+
+
+
+
+
+
+### [4.1.3](https://github.com/furystack/furystack/compare/@furystack/rest-service@4.1.2...@furystack/rest-service@4.1.3) (2021-11-29)
+
+**Note:** Version bump only for package @furystack/rest-service
+
+
+
+
+
+
+### [4.1.2](https://github.com/furystack/furystack/compare/@furystack/rest-service@4.1.1...@furystack/rest-service@4.1.2) (2021-11-19)
+
+**Note:** Version bump only for package @furystack/rest-service
+
+
+
+
+
+
+### [4.1.1](https://github.com/furystack/furystack/compare/@furystack/rest-service@4.1.0...@furystack/rest-service@4.1.1) (2021-11-17)
+
+**Note:** Version bump only for package @furystack/rest-service
+
+
+
+
+
+
+## [4.1.0](https://github.com/furystack/furystack/compare/@furystack/rest-service@4.0.21...@furystack/rest-service@4.1.0) (2021-11-09)
+
+
+### 🚀 What's new
+
+* **@furystack/rest-service:** disabled strict mode for schema validation ([20a9bad](https://github.com/furystack/furystack/commit/20a9bad9368b06b6b3186273b767dbb690e6473a))
+
+
+
+
+### [4.0.21](https://github.com/furystack/furystack/compare/@furystack/rest-service@4.0.20...@furystack/rest-service@4.0.21) (2021-10-15)
+
+**Note:** Version bump only for package @furystack/rest-service
+
+
+
+
+
+
+### [4.0.20](https://github.com/furystack/furystack/compare/@furystack/rest-service@4.0.19...@furystack/rest-service@4.0.20) (2021-10-05)
+
+**Note:** Version bump only for package @furystack/rest-service
+
+
+
+
+
+
+### [4.0.19](https://github.com/furystack/furystack/compare/@furystack/rest-service@4.0.18...@furystack/rest-service@4.0.19) (2021-09-16)
+
+**Note:** Version bump only for package @furystack/rest-service
+
+
+
+
+
+
+### [4.0.18](https://github.com/furystack/furystack/compare/@furystack/rest-service@4.0.17...@furystack/rest-service@4.0.18) (2021-08-27)
+
+**Note:** Version bump only for package @furystack/rest-service
+
+
+
+
+
+
+### [4.0.17](https://github.com/furystack/furystack/compare/@furystack/rest-service@4.0.16...@furystack/rest-service@4.0.17) (2021-08-25)
+
+**Note:** Version bump only for package @furystack/rest-service
+
+
+
+
+
+
 ### [4.0.16](https://github.com/furystack/furystack/compare/@furystack/rest-service@4.0.15...@furystack/rest-service@4.0.16) (2021-08-19)
 
 **Note:** Version bump only for package @furystack/rest-service