+++ packages/rest/CHANGELOG.md
@@ -3,6 +3,79 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [3.1.2](https://github.com/furystack/furystack/compare/@furystack/rest@3.1.1...@furystack/rest@3.1.2) (2021-11-30)
+
+**Note:** Version bump only for package @furystack/rest
+
+
+
+
+
+
+### [3.1.1](https://github.com/furystack/furystack/compare/@furystack/rest@3.1.0...@furystack/rest@3.1.1) (2021-11-29)
+
+**Note:** Version bump only for package @furystack/rest
+
+
+
+
+
+
+## [3.1.0](https://github.com/furystack/furystack/compare/@furystack/rest@3.0.21...@furystack/rest@3.1.0) (2021-11-19)
+
+
+### 🚀 What's new
+
+* **@furystack/rest:** Extended allowed HTTP methods with HEAD, CONNECT and TRACE ([85b91f2](https://github.com/furystack/furystack/commit/85b91f2254a7459d4cb121caeccec792880467fa))
+
+
+
+
+### [3.0.21](https://github.com/furystack/furystack/compare/@furystack/rest@3.0.20...@furystack/rest@3.0.21) (2021-11-09)
+
+**Note:** Version bump only for package @furystack/rest
+
+
+
+
+
+
+### [3.0.20](https://github.com/furystack/furystack/compare/@furystack/rest@3.0.19...@furystack/rest@3.0.20) (2021-10-15)
+
+**Note:** Version bump only for package @furystack/rest
+
+
+
+
+
+
+### [3.0.19](https://github.com/furystack/furystack/compare/@furystack/rest@3.0.18...@furystack/rest@3.0.19) (2021-10-05)
+
+**Note:** Version bump only for package @furystack/rest
+
+
+
+
+
+
+### [3.0.18](https://github.com/furystack/furystack/compare/@furystack/rest@3.0.17...@furystack/rest@3.0.18) (2021-09-16)
+
+**Note:** Version bump only for package @furystack/rest
+
+
+
+
+
+
+### [3.0.17](https://github.com/furystack/furystack/compare/@furystack/rest@3.0.16...@furystack/rest@3.0.17) (2021-08-27)
+
+**Note:** Version bump only for package @furystack/rest
+
+
+
+
+
+
 ### [3.0.16](https://github.com/furystack/furystack/compare/@furystack/rest@3.0.15...@furystack/rest@3.0.16) (2021-08-19)
 
 **Note:** Version bump only for package @furystack/rest