+++ packages/sequelize-store/CHANGELOG.md
@@ -3,6 +3,78 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.3.16](https://github.com/furystack/core/compare/@furystack/sequelize-store@0.3.15...@furystack/sequelize-store@0.3.16) (2021-11-30)
+
+**Note:** Version bump only for package @furystack/sequelize-store
+
+
+
+
+
+
+### [0.3.15](https://github.com/furystack/core/compare/@furystack/sequelize-store@0.3.14...@furystack/sequelize-store@0.3.15) (2021-11-29)
+
+**Note:** Version bump only for package @furystack/sequelize-store
+
+
+
+
+
+
+### [0.3.14](https://github.com/furystack/core/compare/@furystack/sequelize-store@0.3.13...@furystack/sequelize-store@0.3.14) (2021-11-19)
+
+**Note:** Version bump only for package @furystack/sequelize-store
+
+
+
+
+
+
+### [0.3.13](https://github.com/furystack/core/compare/@furystack/sequelize-store@0.3.12...@furystack/sequelize-store@0.3.13) (2021-11-09)
+
+**Note:** Version bump only for package @furystack/sequelize-store
+
+
+
+
+
+
+### [0.3.12](https://github.com/furystack/core/compare/@furystack/sequelize-store@0.3.11...@furystack/sequelize-store@0.3.12) (2021-10-15)
+
+**Note:** Version bump only for package @furystack/sequelize-store
+
+
+
+
+
+
+### [0.3.11](https://github.com/furystack/core/compare/@furystack/sequelize-store@0.3.10...@furystack/sequelize-store@0.3.11) (2021-10-05)
+
+**Note:** Version bump only for package @furystack/sequelize-store
+
+
+
+
+
+
+### [0.3.10](https://github.com/furystack/core/compare/@furystack/sequelize-store@0.3.9...@furystack/sequelize-store@0.3.10) (2021-09-16)
+
+**Note:** Version bump only for package @furystack/sequelize-store
+
+
+
+
+
+
+### [0.3.9](https://github.com/furystack/core/compare/@furystack/sequelize-store@0.3.8...@furystack/sequelize-store@0.3.9) (2021-08-27)
+
+**Note:** Version bump only for package @furystack/sequelize-store
+
+
+
+
+
+
 ### [0.3.8](https://github.com/furystack/core/compare/@furystack/sequelize-store@0.3.7...@furystack/sequelize-store@0.3.8) (2021-08-19)
 
 **Note:** Version bump only for package @furystack/sequelize-store