+++ packages/shades/CHANGELOG.md
@@ -3,6 +3,88 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [3.7.3](https://github.com/furystack/furystack/compare/@furystack/shades@3.7.2...@furystack/shades@3.7.3) (2021-11-30)
+
+**Note:** Version bump only for package @furystack/shades
+
+
+
+
+
+
+### [3.7.2](https://github.com/furystack/furystack/compare/@furystack/shades@3.7.1...@furystack/shades@3.7.2) (2021-11-29)
+
+**Note:** Version bump only for package @furystack/shades
+
+
+
+
+
+
+### [3.7.1](https://github.com/furystack/furystack/compare/@furystack/shades@3.7.0...@furystack/shades@3.7.1) (2021-11-19)
+
+**Note:** Version bump only for package @furystack/shades
+
+
+
+
+
+
+## [3.7.0](https://github.com/furystack/furystack/compare/@furystack/shades@3.6.6...@furystack/shades@3.7.0) (2021-11-17)
+
+
+### 🚀 What's new
+
+* **@furystack/shades:** Attach Data attributes to native elements ([6225bcb](https://github.com/furystack/furystack/commit/6225bcb0cb6fd311f6761317a86d4d0e5b30bb3e))
+
+
+
+
+### [3.6.6](https://github.com/furystack/furystack/compare/@furystack/shades@3.6.5...@furystack/shades@3.6.6) (2021-11-09)
+
+**Note:** Version bump only for package @furystack/shades
+
+
+
+
+
+
+### [3.6.5](https://github.com/furystack/furystack/compare/@furystack/shades@3.6.4...@furystack/shades@3.6.5) (2021-10-15)
+
+**Note:** Version bump only for package @furystack/shades
+
+
+
+
+
+
+### [3.6.4](https://github.com/furystack/furystack/compare/@furystack/shades@3.6.3...@furystack/shades@3.6.4) (2021-10-05)
+
+**Note:** Version bump only for package @furystack/shades
+
+
+
+
+
+
+### [3.6.3](https://github.com/furystack/furystack/compare/@furystack/shades@3.6.2...@furystack/shades@3.6.3) (2021-09-16)
+
+**Note:** Version bump only for package @furystack/shades
+
+
+
+
+
+
+### [3.6.2](https://github.com/furystack/furystack/compare/@furystack/shades@3.6.1...@furystack/shades@3.6.2) (2021-08-27)
+
+**Note:** Version bump only for package @furystack/shades
+
+
+
+
+
+
 ### [3.6.1](https://github.com/furystack/furystack/compare/@furystack/shades@3.6.0...@furystack/shades@3.6.1) (2021-08-19)
 
 **Note:** Version bump only for package @furystack/shades