+++ packages/utils/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [1.2.46](https://github.com/furystack/furystack/compare/@furystack/utils@1.2.45...@furystack/utils@1.2.46) (2021-11-19)
+
+**Note:** Version bump only for package @furystack/utils
+
+
+
+
+
+
+### [1.2.45](https://github.com/furystack/furystack/compare/@furystack/utils@1.2.44...@furystack/utils@1.2.45) (2021-10-05)
+
+**Note:** Version bump only for package @furystack/utils
+
+
+
+
+
+
 ### [1.2.44](https://github.com/furystack/furystack/compare/@furystack/utils@1.2.14...@furystack/utils@1.2.44) (2021-08-19)
 
 **Note:** Version bump only for package @furystack/utils