+++ packages/websocket-api/CHANGELOG.md
@@ -3,6 +3,96 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [7.0.56](https://github.com/furystack/furystack/compare/@furystack/websocket-api@7.0.55...@furystack/websocket-api@7.0.56) (2021-11-30)
+
+**Note:** Version bump only for package @furystack/websocket-api
+
+
+
+
+
+
+### [7.0.55](https://github.com/furystack/furystack/compare/@furystack/websocket-api@7.0.54...@furystack/websocket-api@7.0.55) (2021-11-29)
+
+**Note:** Version bump only for package @furystack/websocket-api
+
+
+
+
+
+
+### [7.0.54](https://github.com/furystack/furystack/compare/@furystack/websocket-api@7.0.53...@furystack/websocket-api@7.0.54) (2021-11-19)
+
+**Note:** Version bump only for package @furystack/websocket-api
+
+
+
+
+
+
+### [7.0.53](https://github.com/furystack/furystack/compare/@furystack/websocket-api@7.0.52...@furystack/websocket-api@7.0.53) (2021-11-17)
+
+**Note:** Version bump only for package @furystack/websocket-api
+
+
+
+
+
+
+### [7.0.52](https://github.com/furystack/furystack/compare/@furystack/websocket-api@7.0.51...@furystack/websocket-api@7.0.52) (2021-11-09)
+
+**Note:** Version bump only for package @furystack/websocket-api
+
+
+
+
+
+
+### [7.0.51](https://github.com/furystack/furystack/compare/@furystack/websocket-api@7.0.50...@furystack/websocket-api@7.0.51) (2021-10-15)
+
+**Note:** Version bump only for package @furystack/websocket-api
+
+
+
+
+
+
+### [7.0.50](https://github.com/furystack/furystack/compare/@furystack/websocket-api@7.0.49...@furystack/websocket-api@7.0.50) (2021-10-05)
+
+**Note:** Version bump only for package @furystack/websocket-api
+
+
+
+
+
+
+### [7.0.49](https://github.com/furystack/furystack/compare/@furystack/websocket-api@7.0.48...@furystack/websocket-api@7.0.49) (2021-09-16)
+
+**Note:** Version bump only for package @furystack/websocket-api
+
+
+
+
+
+
+### [7.0.48](https://github.com/furystack/furystack/compare/@furystack/websocket-api@7.0.47...@furystack/websocket-api@7.0.48) (2021-08-27)
+
+**Note:** Version bump only for package @furystack/websocket-api
+
+
+
+
+
+
+### [7.0.47](https://github.com/furystack/furystack/compare/@furystack/websocket-api@7.0.46...@furystack/websocket-api@7.0.47) (2021-08-25)
+
+**Note:** Version bump only for package @furystack/websocket-api
+
+
+
+
+
+
 ### [7.0.46](https://github.com/furystack/furystack/compare/@furystack/websocket-api@7.0.45...@furystack/websocket-api@7.0.46) (2021-08-19)
 
 **Note:** Version bump only for package @furystack/websocket-api